Swap to a pod_template capturing all the complex crap
Kustomize can rewrite configs generated using config map and secret
generators correctly when specified via a PodTemplate. This allows
existing tooling to be used to specify whatever should be run at some
point, but the controller can own spawning of a job with generateName
which kustomize cannot do. This seems to feel a little bit better.